requirements.txt
tensorflow>=2.16.0
numpy>=1.26.0
pillow>=10.0.0
matplotlib>=3.8.0
scikit-learn>=1.4.0
gradio>=4.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
